pipelines/dagster/requirements.txt
# Dagster core
dagster==1.5.10
dagster-webserver==1.5.10
dagster-postgres==0.21.10
dagster-graphql==1.5.10

# Google Cloud
google-cloud-storage
google-cloud-pubsub

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
pandas==2.1.3
numpy==1.26.2

# AI/ML
openai>=1.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
